Merge UI branch

+++ Code/Platforms/OpenGL/OpenGLTexture.h
@@ -18,7 +18,7 @@ public:
 
 	virtual int GetWidth() const override { return m_width; }
 	virtual int GetHeight() const override { return m_height; }
-	virtual uint32_t GetRendererID() const override { return m_id; }
+	virtual uint64_t GetRendererID() const override { return m_id; }
 
 private:
 	GLuint m_id;
